requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
sqlalchemy==2.0.40
alembic==1.15.2
pydantic[email]==2.11.4
nanoid==2.0.0
PyJWT==2.8.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
bcrypt==4.2.0